+++ scripts/build-extension.mjs
@@ -16,13 +16,13 @@ const argv = await parseArgs(process.argv);
 /**
  * Main build function.
  *
- * @param {string}  platform The platform to build for.
- * @param {boolean} watch    Whether to watch for changes.
- * @param {string}  release  Whether building for release.
+ * @param {string}  platform  The platform to build for.
+ * @param {boolean} watch     Whether to watch for changes.
+ * @param {boolean} isRelease Whether this is a release build.
  *
  * @return {Promise<void>} Esbuild build result.
  */
-const buildExtension = async (platform, watch = false, release = false) => {
+const buildExtension = async (platform, watch = false, isRelease = false) => {
   fs.mkdirSync(path.join(process.cwd(), `dist/${platform}`), { recursive: true });
 
   // Copy manifest.json and inject the version number.
@@ -57,6 +57,7 @@ const buildExtension = async (platform, watch = false, release = false) => {
     alias: {
       '@data': path.resolve(process.cwd(), 'dist/data'),
     },
+    dropLabels: ['excludeFromExtension'],
     outfile: `dist/${platform}/main.js`,
     plugins: [
       ImportGlobPlugin,
@@ -90,12 +91,12 @@ const buildExtension = async (platform, watch = false, release = false) => {
       ].join('\n'),
     },
     define: {
-      __SENTRY_DSN__: JSON.stringify(release ? 'https://a677b0fe4d2fbc3a7db7410353d91f39@o4506582061875200.ingest.sentry.io/4506781071835136' : ''),
+      __SENTRY_DSN__: JSON.stringify(isRelease ? 'https://a677b0fe4d2fbc3a7db7410353d91f39@o4506582061875200.ingest.sentry.io/4506781071835136' : ''),
     }
   };
 
-  console.log(watch ? 'Watching for changes...' : 'Building extension...'); // eslint-disable-line no-console
   if (watch) {
+    console.log(' Watching for changes...'); // eslint-disable-line no-console
     opts.logLevel = 'info';
     const ctx = await esbuild.context(opts);
     return await ctx.watch();