+++ src/modules/better-maps/index.js
@@ -288,7 +288,7 @@ const updateRelicHunterHint = async () => {
 };
 
 let _showInventory;
-const relicHunterUpdate = () => {
+const relicHunterUpdate = async () => {
   if (_showInventory) {
     return;
   }
@@ -298,17 +298,19 @@ const relicHunterUpdate = () => {
   /**
    * Show the inventory.
    */
-  hg.controllers.TreasureMapController.showInventory = () => {
+  hg.controllers.TreasureMapController.showInventory = async () => {
     _showInventory();
 
     // Call updateRelicHunterHint, but if it fails, try again in 250ms, but stop after 5 tries.
-    let tries = 0;
-    const interval = setInterval(async () => {
-      tries++;
-      if (await updateRelicHunterHint() || tries > 5) {
-        clearInterval(interval);
-      }
-    }, 250);
+    if (! await updateRelicHunterHint()) {
+      let tries = 0;
+      const interval = setInterval(async () => {
+        tries++;
+        if (await updateRelicHunterHint() || tries > 5) {
+          clearInterval(interval);
+        }
+      }, 250);
+    }
   };
 };